telegram bot/requirements.txt
pyTelegramBotAPI
python-dotenv
inference-sdk 
nest_asyncio 